Rename StreamClient to StreamSender
